move bubble styles into BubbleFactory
