+++ apps/src/templates/progress/BubbleFactory.jsx
@@ -5,17 +5,20 @@ import classNames from 'classnames';
 import _ from 'lodash';
 import queryString from 'query-string';
 import i18n from '@cdo/locale';
-import {currentLocation} from '@cdo/apps/utils';
+import {currentLocation, makeEnum} from '@cdo/apps/utils';
 import TooltipWithIcon from './TooltipWithIcon';
 import {getIconForLevel, isLevelAssessment} from './progressHelpers';
 import {
-  bubbleStyles,
-  BubbleSize,
-  BubbleShape,
-  mainBubbleStyle,
-  diamondContainerStyle
+  flex,
+  font,
+  marginLeftRight,
+  marginTopBottom,
+  tightlyConstrainedSizeStyle
 } from './progressStyles';
 import {levelWithProgressType} from './progressTypes';
+
+export const BubbleSize = makeEnum('dot', 'letter', 'full');
+export const BubbleShape = makeEnum('circle', 'diamond', 'pill');
 
 /**
  * Base bubble component defined in terms of shape, size, and style, as opposed
@@ -195,6 +198,135 @@ export function getBubbleUrl(
   return levelUrl;
 }
 
+/**
+ * ======================================
+ * Bubble styles
+ * ======================================
+ */
+
+const LARGE_FONT = 16;
+const SMALL_FONT = 12;
+
+const bubbleSizes = {
+  [BubbleShape.circle]: {
+    [BubbleSize.dot]: 13,
+    [BubbleSize.letter]: 20,
+    [BubbleSize.full]: 34
+  },
+  [BubbleShape.diamond]: {
+    [BubbleSize.dot]: 10,
+    [BubbleSize.full]: 26
+  },
+  [BubbleShape.pill]: {}
+};
+
+const circleMargins = {
+  [BubbleSize.dot]: 3,
+  [BubbleSize.letter]: 3,
+  [BubbleSize.full]: 2
+};
+
+const bubbleBorderRadii = {
+  [BubbleShape.circle]: {
+    [BubbleSize.dot]: bubbleSizes[BubbleShape.circle][BubbleSize.dot],
+    [BubbleSize.letter]: bubbleSizes[BubbleShape.circle][BubbleSize.letter],
+    [BubbleSize.full]: bubbleSizes[BubbleShape.circle][BubbleSize.full]
+  },
+  [BubbleShape.diamond]: {
+    [BubbleSize.dot]: 2,
+    [BubbleSize.full]: 4
+  },
+  [BubbleShape.pill]: {}
+};
+
+/**
+ * We use fixed-size containers to make diamond bubbles the same width as
+ * circle bubbles, so we use BubbleShape.circle to compute.
+ *
+ * Container width is the width of the bubble plus the left and right margins.
+ */
+export const bubbleContainerWidths = {
+  [BubbleSize.dot]:
+    bubbleSizes[BubbleShape.circle][BubbleSize.dot] +
+    2 * circleMargins[BubbleSize.dot],
+  [BubbleSize.letter]:
+    bubbleSizes[BubbleShape.circle][BubbleSize.letter] +
+    2 * circleMargins[BubbleSize.letter],
+  [BubbleSize.full]:
+    bubbleSizes[BubbleShape.circle][BubbleSize.full] +
+    2 * circleMargins[BubbleSize.full]
+};
+
+const fontSizes = {
+  [BubbleSize.letter]: SMALL_FONT,
+  [BubbleSize.full]: LARGE_FONT
+};
+
+const bubbleStyles = {
+  main: {
+    ...flex,
+    ...font,
+    ...marginTopBottom(3),
+    boxSizing: 'border-box',
+    letterSpacing: -0.11,
+    position: 'relative',
+    whiteSpace: 'nowrap'
+  },
+  pill: {
+    borderRadius: 20,
+    fontSize: SMALL_FONT,
+    padding: '6px 10px'
+  },
+  diamond: {
+    ...marginTopBottom(6),
+    transform: 'rotate(45deg)',
+    padding: 2
+  },
+  diamondContentTransform: {
+    transform: 'rotate(-45deg)'
+  }
+};
+
+/**
+ * Computes style for shape/size, and merges with `progressStyle` previously
+ * computed by `progressStyles.levelProgressStyle`.
+ */
+function mainBubbleStyle(shape, size, progressStyle) {
+  return {
+    ...bubbleStyles.main,
+    ...shapeSizeStyle(shape, size),
+    ...progressStyle
+  };
+}
+
+function diamondContainerStyle(size) {
+  const containerWidth = bubbleContainerWidths[size];
+  return {
+    ...flex,
+    width: containerWidth,
+    height: containerWidth
+  };
+}
+
+function shapeSizeStyle(shape, size) {
+  if (shape === BubbleShape.pill) {
+    return bubbleStyles.pill;
+  }
+
+  const bubbleSize = bubbleSizes[shape][size];
+  const fontSize = fontSizes[size];
+  return {
+    ...tightlyConstrainedSizeStyle(bubbleSize),
+    borderRadius: bubbleBorderRadii[shape][size],
+    fontSize: fontSize,
+    lineHeight: `${fontSize}px`,
+    ...(shape === BubbleShape.circle && marginLeftRight(circleMargins[size])),
+    ...(shape === BubbleShape.diamond && bubbleStyles.diamond)
+  };
+}
+
 export const unitTestExports = {
-  DiamondContainer
+  DiamondContainer,
+  bubbleStyles,
+  mainBubbleStyle
 };
