+++ apps/src/templates/progress/progressStyles.js
@@ -1,7 +1,15 @@
 import color from '@cdo/apps/util/color';
-import {makeEnum} from '@cdo/apps/utils';
 import {LevelStatus, LevelKind} from '@cdo/apps/util/sharedConstants';
 import styles from './styles.scss';
+
+/**
+ * Note: these constants will be removed in favor of `BubbleFactory.bubbleSizes`
+ * once we finish cleaning up all of our bubble components (LP-1662).
+ */
+export const DOT_SIZE = 30;
+export const DIAMOND_DOT_SIZE = 22;
+export const SMALL_DOT_SIZE = 9;
+export const SMALL_DIAMOND_SIZE = 6;
 
 /**
  * ======================================
@@ -76,145 +84,6 @@ export const hoverStyle = {
   transition:
     'background-color .2s ease-out, border-color .2s ease-out, color .2s ease-out'
 };
-
-/**
- * ======================================
- * Bubble styles
- * ======================================
- */
-
-export const BubbleSize = makeEnum('dot', 'letter', 'full');
-export const BubbleShape = makeEnum('circle', 'diamond', 'pill');
-
-/**
- * Note: these constants will be removed in favor of `bubbleSizes` below once
- * we finish cleaning up all of our bubble components.
- */
-export const DOT_SIZE = 30;
-export const DIAMOND_DOT_SIZE = 22;
-export const SMALL_DOT_SIZE = 9;
-export const SMALL_DIAMOND_SIZE = 6;
-
-const LARGE_FONT = 16;
-const SMALL_FONT = 12;
-
-const bubbleSizes = {
-  [BubbleShape.circle]: {
-    [BubbleSize.dot]: 13,
-    [BubbleSize.letter]: 20,
-    [BubbleSize.full]: 34
-  },
-  [BubbleShape.diamond]: {
-    [BubbleSize.dot]: 10,
-    [BubbleSize.full]: 26
-  },
-  [BubbleShape.pill]: {}
-};
-
-const circleMargins = {
-  [BubbleSize.dot]: 3,
-  [BubbleSize.letter]: 3,
-  [BubbleSize.full]: 2
-};
-
-const bubbleBorderRadii = {
-  [BubbleShape.circle]: {
-    [BubbleSize.dot]: bubbleSizes[BubbleShape.circle][BubbleSize.dot],
-    [BubbleSize.letter]: bubbleSizes[BubbleShape.circle][BubbleSize.letter],
-    [BubbleSize.full]: bubbleSizes[BubbleShape.circle][BubbleSize.full]
-  },
-  [BubbleShape.diamond]: {
-    [BubbleSize.dot]: 2,
-    [BubbleSize.full]: 4
-  },
-  [BubbleShape.pill]: {}
-};
-
-/**
- * We use fixed-size containers to make diamond bubbles the same width as
- * circle bubbles, so we use BubbleShape.circle to compute.
- *
- * Container width is the width of the bubble plus the left and right margins.
- */
-export const bubbleContainerWidths = {
-  [BubbleSize.dot]:
-    bubbleSizes[BubbleShape.circle][BubbleSize.dot] +
-    2 * circleMargins[BubbleSize.dot],
-  [BubbleSize.letter]:
-    bubbleSizes[BubbleShape.circle][BubbleSize.letter] +
-    2 * circleMargins[BubbleSize.letter],
-  [BubbleSize.full]:
-    bubbleSizes[BubbleShape.circle][BubbleSize.full] +
-    2 * circleMargins[BubbleSize.full]
-};
-
-const fontSizes = {
-  [BubbleSize.letter]: SMALL_FONT,
-  [BubbleSize.full]: LARGE_FONT
-};
-
-export const bubbleStyles = {
-  main: {
-    ...flex,
-    ...font,
-    ...marginTopBottom(3),
-    boxSizing: 'border-box',
-    letterSpacing: -0.11,
-    position: 'relative',
-    whiteSpace: 'nowrap'
-  },
-  pill: {
-    borderRadius: 20,
-    fontSize: SMALL_FONT,
-    padding: '6px 10px'
-  },
-  diamond: {
-    ...marginTopBottom(6),
-    transform: 'rotate(45deg)',
-    padding: 2
-  },
-  diamondContentTransform: {
-    transform: 'rotate(-45deg)'
-  }
-};
-
-/**
- * Computes style for shape/size, and merges with `progressStyle` previously
- * computed by `levelProgressStyle` below.
- */
-export function mainBubbleStyle(shape, size, progressStyle) {
-  return {
-    ...bubbleStyles.main,
-    ...shapeSizeStyle(shape, size),
-    ...progressStyle
-  };
-}
-
-export function diamondContainerStyle(size) {
-  const containerWidth = bubbleContainerWidths[size];
-  return {
-    ...flex,
-    width: containerWidth,
-    height: containerWidth
-  };
-}
-
-function shapeSizeStyle(shape, size) {
-  if (shape === BubbleShape.pill) {
-    return bubbleStyles.pill;
-  }
-
-  const bubbleSize = bubbleSizes[shape][size];
-  const fontSize = fontSizes[size];
-  return {
-    ...tightlyConstrainedSizeStyle(bubbleSize),
-    borderRadius: bubbleBorderRadii[shape][size],
-    fontSize: fontSize,
-    lineHeight: `${fontSize}px`,
-    ...(shape === BubbleShape.circle && marginLeftRight(circleMargins[size])),
-    ...(shape === BubbleShape.diamond && bubbleStyles.diamond)
-  };
-}
 
 /**
  * ======================================
